pyproject.toml
[project]
name = "google-maps-ai-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain-community>=0.3.26",
    "langchain-core>=0.3.66",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.0",
    "loguru>=0.7.3",
    "openai>=1.93.0",
    "openai-agents>=0.1.0",
    "playwright>=1.53.0",
    "pydantic>=2.11.7",
    "pyotp>=2.9.0",
    "python-dotenv>=1.1.1",
    "uuid>=1.30",
]
